volta/requirements.txt
Counter==1.0.0
DateTime==3.0.2
Django==1.4.3
PyYAML==3.10
South==0.7.6
distribute==0.6.31
dj-database-url==0.2.1
django-classy-tags==0.3.4.1
django-cms==2.3.5
django-mptt==0.5.2
django-reversion==1.6.5
django-sekizai==0.6.1
html5lib==0.95
networkx==1.7
nltk==2.0.4
ordereddict==1.1
pytz==2012j
zope.interface==4.0.3
PIL==1.1.7
psycopg2==2.4.6
skype4py==1.0.32.0
xlrd==0.7.7
xlwt==0.7.4
